requirements.txt
allure-pytest==2.13.0
allure-python-commons==2.13.0
pytest==7.2.2
requests==2.28.2
selenium==4.8.2
webdriver-manager==3.8.5
jsonpath~=0.82